apps/worker-api/pyproject.toml
[tool.poetry]
name = "worker-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
azure-data-tables = "^12.5.0"
pygithub = "^2.4.0"
aiohappyeyeballs = "^2.4.0"
aiohttp = "^3.10.5"
aiosignal = "^1.3.1"
annotated-types = "^0.7.0"
anyio = "^4.4.0"
attrs = "^24.2.0"
azure-common = "^1.1.28"
azure-core = "^1.31.0"
azure-identity = "^1.17.1"
azure-search-documents = "^11.5.1"
beautifulsoup4 = "^4.12.3"
certifi = "^2024.8.30"
cffi = "^1.17.1"
chardet = "^5.2.0"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
cryptography = "^43.0.1"
dataclasses-json = "^0.6.7"
deprecated = "^1.2.14"
distro = "^1.9.0"
frozenlist = "^1.4.1"
greenlet = "^3.1.0"
h11 = "^0.14.0"
httpcore = "^1.0.5"
httpx = "^0.27.2"
idna = "^3.10"
isodate = "^0.6.1"
jiter = "^0.5.0"
jsonpatch = "^1.33"
jsonpointer = "^3.0.0"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-core = "^0.3.1"
langchain-openai = "^0.2.0"
langchain-text-splitters = "^0.3.0"
langsmith = "^0.1.121"
marshmallow = "^3.22.0"
msal = "^1.31.0"
msal-extensions = "^1.2.0"
multidict = "^6.1.0"
mypy-extensions = "^1.0.0"
numpy = "^1.26.4"
openai = "^1.46.0"
orjson = "^3.10.7"
packaging = "^24.1"
portalocker = "^2.10.1"
pycparser = "^2.22"
pydantic = "^2.9.2"
pydantic-core = "^2.23.4"
pydantic-settings = "^2.5.2"
pyjwt = { version = "^2.9.0", extras = ["crypto"] }
pynacl = "^1.5.0"
pyyaml = "^6.0.2"
regex = "^2024.9.11"
six = "^1.16.0"
sniffio = "^1.3.1"
soupsieve = "^2.6"
sqlalchemy = "^2.0.35"
starlette = "^0.38.5"
tenacity = "^8.5.0"
tiktoken = "^0.7.0"
tqdm = "^4.66.5"
typing-extensions = "^4.12.2"
typing-inspect = "^0.9.0"
urllib3 = "^2.2.3"
wrapt = "^1.16.0"
yarl = "^1.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
